requirements.txt
flask
python-dotenv # Debug mode: on
psycopg2